Data.Algebra.Free

Description

Author : Marcin Szamotulski, 2018 Copyright : (c) Marcin Szamotulski, 2018 License : MPL 2.0 (https:/www.mozilla.orgen-USMPL2.0/) Stability : Experimental

This is a GHC implementation of Universal Algebra approach to free algebras. For a theoretical introduction to Universal Algebra and free algebras please read "From Free Algebras to Free Monads".

Algebra type and free algebras

type family AlgebraType (a :: k) (b :: l) :: Constraint

We will use this type class to define constrains of concrete algebras, e.g.

type instance AlgebraType [a] m = Monoid m
type instance AlgebraType FreeCommutativeMonoid m = (Monoid m, CommutativeSemigroup m)

class FreeAlgebra (m :: * -> *) where

A free algebra of type AlgebraType. A FreeAlgebra constraint proves that m is a free algebra in the class of algebras of type AlgebraType m over the generators gen.

A lawful instance should imply AlgebraType m (m a) for all a.

Minimal complete definition

gen, free

Methods

gen :: a -> m a

Injective map that embeds generators of m.

free

Arguments

:: AlgebraType m d

=> (a -> d)
map generators of m into d

-> m a -> d
returns a homomorphism from m to d

The freeness property.

foldFree :: (FreeAlgebra m, AlgebraType m a) => m a -> a

All types which satisfy FreeAlgebra constraint are foldable.

hoistFree :: forall m n a. (AlgebraType m (n a), FreeAlgebra m, FreeAlgebra n) => m a -> n a

The canonical quotient map from a free algebra of a wider class to a free algebra of a narrower class, e.g. from free non-associative semigroup to free associative semigroup, free monoid to free commutative monoid, etc.

hoistFree . hoistFree == hoistFree
mapFree f . hoistFree == hiostFree . mapFree f

mapFree :: forall m a b. (FreeAlgebra m, AlgebraType m (m b)) => (a -> b) -> m a -> m b

All types which satisfy FreeAlgebra constraint are functors. The constraint AlgebraType m (m b) is always satisfied.

joinFree :: (FreeAlgebra m, AlgebraType m (m a)) => m (m a) -> m a

FreeAlgebra constraint imply monad.

bindFree :: (FreeAlgebra m, AlgebraType m (m b), AlgebraType m (m (m b))) => m a -> (a -> m b) -> m b

The monadic bind operator.

Non-Associative Semigroups

class NASemigroup a where

Non-Associative (abbrev NA) semigroup bears <> method which is not subject to any laws.

Minimal complete definition

(<>)

Methods

(<>) :: a -> a -> a

data FreeNASemigroup a

Free non-associative semigroup generated by a type a.

Constructors

Generator a

(FreeNASemigroup a) :*: (FreeNASemigroup a)

foldNA :: NASemigroup a => FreeNASemigroup a -> a

newtype Conj

Constructors

Conj

Fields

runConj :: Bool

newtype Disj

Constructors

Disj

Fields

runDisj :: Bool

Examples

type FreeNASemigroup1 = FreeNASemigroup ()

Non associative natural numbers:

(), () :*: (), (() :*: ()) :*: (), () :*: (() :*: ()), ...

type FreeNASemigroup2 = FreeNASemigroup (Sum (Succ (Succ Zero)) ())

Free non associative semigroup with two generators.

Associative Semigroups

class NASemigroup m => Semigroup m

Class for associative semigroups, for which:

 (a <> b) <> c = a <> (b <> c)

newtype FreeSemigroup a

Free semigroup

Constructors

FreeSemigroup

Fields

runFreeSemigroup :: NonEmpty a

foldNonEmpty :: Semigroup a => FreeSemigroup a -> a

type FreeSemigroup1 = FreeSemigroup ()

Free (associative) semigroup with one generator, which is commutative.

type FreeSemigroup2 = FreeSemigroup (Either () ())

Free (associative) semigroup with two generators, hence non commutative.

toAssociative :: forall a. FreeNASemigroup a -> FreeSemigroup a

This map is the canonical epimorphism from the free non-associative semigroup to free associative semigroup. It is uniquely determined on generators.

>>> toAssociative $ Generator () :*: (Generator () :*: Generator ())
FreeSemigroup () :| [(),()]
>>> toAssociative $ (Generator () :*: Generator ()) :*: Generator ()
FreeSemigroup () :| [(),()]

>>> toAssociative (mapFree (\_ -> 0) (Generator () :*: Generator ()))
FreeSemigroup 0 :| [0]
>>> mapFree (\_ -> 0) $ toAssociative (Generator () :*: Generator ())
FreeSemigroup 0 :| [0]

Pointed Algebras

class Pointed m where

A class for pointed types, aka pointed sets, i.e. types with a distinguished term. This is one of the simple algebras. Homomorphism of pointed algebras are required to preserve the distinguished element.

Minimal complete definition

mempty

Methods

mempty

Arguments

:: m
distinguished term

Sums and Products

data Nat

Type level natural numbers promoted with DataKinds.

Constructors

Zero

Succ Nat

data Prod (n :: Nat) a where

n-fold products

Constructors

Cons :: a -> Prod n a -> Prod (Succ n) a

EmptyProduct :: Prod Zero a

p1 :: Prod (Succ (Succ (Succ Zero))) [()]

p2 :: Prod (Succ (Succ (Succ Zero))) [()]

p3 :: Prod (Succ (Succ (Succ Zero))) [()]

data Sum (n :: Nat) a where

n-fold sum

Constructors

EmptySum :: Sum Zero a

Add :: Either (Sum n a) a -> Sum (Succ n) a

s1 :: Sum (Succ n) ()

s2 :: Sum (Succ (Succ n)) ()

s3 :: Sum (Succ (Succ (Succ n))) ()

class ToProduct (n :: Nat) m a b where

In Category Theory there is a unique morphism from a (disjoint) sum `a ∪ b` to a product `a × b`. This type class let us construct it inductively.

>>> toProduct @('Succ ('Succ 'Zero)) Just $ Add (Right 0)
Cons (Just 0) (Cons Nothing EmptyProduct) 
>>> toProduct @('Succ ('Succ 'Zero)) Just $ Add (Left (Add $ Right 0))
Cons Nothing (Cons (Just 0) EmptyProduct)

Minimal complete definition

toProduct

Methods

toProduct

Arguments

:: (a -> b)
how to transform coeffcients

-> m n a
a sum type

-> Prod n b
the resulting product

diagonal :: Nat -> a -> (forall n. Prod n a -> r) -> r

Construct any Prod with all values equal in CPS style.

Monoids

class (Semigroup m, Pointed m) => Monoid m

newtype FreeMonoid a

Constructors

FreeMonoid

Fields

runFreeMonoid :: [a]

foldMonoid :: Monoid m => FreeMonoid m -> m

toMonoid :: FreeSemigroup a -> FreeMonoid a

class Semigroup a => CommutativeSemigroup a

The class of commutative semigroups.

a <> b = b <> a

type FreeMonoid1 = FreeMonoid ()

It is isomorphic with Nat (as an additive monoid) via natToFreeMonoid1.

type FreeMonoid2 = FreeMonoid (Sum (Succ (Succ Zero)) ())

type FreeMonoid3 = FreeMonoid (Sum (Succ (Succ (Succ Zero))) ())

f1 :: FreeMonoid3

f2 :: FreeMonoid3

f3 :: FreeMonoid3

Commutative monoids

Free commutative monoid

Higher Kinded Algebras

type family AlgebraType1 (m :: k) (b :: * -> *) :: Constraint

The constraint that will be imposed on the generator type of kind (* -> *).

class FreeAlgebra1 (m :: (* -> *) -> * -> *) where

Higher kinded version of FreeAlgebra. Instances includes free functors, free applicative functors and free monads.

Minimal complete definition

gen1, free1

Methods

gen1 :: AlgebraType1 m f => f a -> m f a

Natural transformation that embeds generators into m.

free1

Arguments

:: forall (d :: * -> *). (AlgebraType m d, AlgebraType1 m f)

=> (forall x. f x -> d x)
natural transformation which embeds generators of m into d

-> (a -> b)

-> m f a -> d b
a homomorphism from m to d

The freeness property.

Note that free1 will allow to build mapFree (and thus joinFree), only because we provide a way to change the type inside d from a to b.

foldFree1

Arguments

:: (FreeAlgebra1 m, AlgebraType m f, AlgebraType1 m f)

=> m f a

-> f a

FreeAlgebra1 m implies that m f is a foldable:

natFree1

Arguments

:: (FreeAlgebra1 m, AlgebraType m (m g), AlgebraType1 m f, AlgebraType1 m g)

=> (forall x. f x -> g x)
a natural transformation f ~> g

-> m f a

-> m g a

This is Functor instance for m when considered as en endofuctor of some subcategory of * -> * (e.g. endofunctors of _Hask_).

hoistFree1

Arguments

:: (AlgebraType m (n f), AlgebraType1 m f, AlgebraType1 n f, FreeAlgebra1 m, FreeAlgebra1 n)

=> m f a

-> n f a

Analogous to hoistFree

hoistFree1 . hoistFree1 = hoistFree1

mapFree1

Arguments

:: (FreeAlgebra1 m, AlgebraType m (m f), AlgebraType1 m f)

=> (a -> b)

-> m f a

-> m f b

FreeAlgegra1 m implies that m f is a functor.

joinFree1

Arguments

:: (FreeAlgebra1 m, AlgebraType m (m f), AlgebraType1 m (m f))

=> m (m f) a

-> m f a

joinFree1 makes m a monad in some subcatgory of types of kind * -> * (usually the end-functor category of Hask).

bindFree1

Arguments

:: (FreeAlgebra1 m, AlgebraType m (m g), AlgebraType m (m (m g)), AlgebraType1 m (m g), AlgebraType1 m f)

=> m f a

-> (forall x. f x -> m g x)
natural transformation f ~> m g

-> m g a

Bind operator for the joinFree1 monad

assoc1 :: forall m f a. (FreeAlgebra1 m, AlgebraType m f, AlgebraType m (m (m f)), AlgebraType1 m f, AlgebraType1 m (m f)) => m f (m f a) -> m (m f) (f a)

Functor1 and Monad1 classes

class Functor1 (m :: (* -> *) -> * -> *) where

Higher order functor, e.g. functor on a (sub)category of *endofuncors*. The usual functor laws::

 fmap1 id = id

 fmap1 g . fmap1 f = fmap1 (f . g)

Minimal complete definition

fmap1

Associated Types

type FunctorType1 m (f :: * -> *) :: Constraint

Methods

fmap1

Arguments

:: (FunctorType1 m f, FunctorType1 m g)

=> (forall x. f x -> g x)
note that by rank n types, one has not access to x, you are supposed to change the functor, e.g. Just . runIdentity :: forall x . Identity x -> Maybe x

-> m f a

-> m g a

class Functor1 m => Monad1 (m :: (* -> *) -> * -> *) where

Higher order monad, e.g. monad in a (sub)category of _endofunctors_. The usaual monad laws apply:

 return1 fa >>=: f = f fa

 mfa >>=: return1  = mfa

 mfa >>=: (\ fa -> k fa >>=: h) = (mfa '>>=: k) >>=: h

Minimal complete definition

return1, (>>=:)

Methods

return1 :: FunctorType1 m f => f a -> m f a

(>>=:) :: (FunctorType1 m f, FunctorType1 m g) => m f a -> (forall x. f x -> m g x) -> m g a

Coyoneda

 data Coyoneda f a = Coyoneda (b -> a) -> f b

It is the free construction for a functor.

foldCoyoneda :: Functor f => Coyoneda f a -> f a

Free applicative functor

data FreeA f a where

Constructors

Pure :: a -> FreeA f a

FreeA :: f (x -> a) -> FreeA f x -> FreeA f a

data FreeAL f a where

Constructors

PureL :: a -> FreeAL f a

FreeAL :: f (x -> a) -> f x -> FreeAL f a

toFreeA :: forall f a. Functor f => FreeAL f a -> FreeA f a

A smart constructor of FreeA. (and an isomorphism).

foldFreeA :: Applicative f => FreeA f a -> f a

>>> foldFreeA (FreeA (Identity (+1)) (Pure 0))
Identity 1

coyonedaToFreeA :: forall f a. Functor f => Coyoneda f a -> FreeA f a

Day convolution

data Day f g a = Day (f b) (g c) (b -> c -> a)

Note that for Day f g to be an applicative functor it is not enough that f and g are functors, one needs to assume that both are applicative. Hence Day f f is not a candidate for a free applicative functor generated by a functor f. Nevertheless it bears FreeAlgebra1 property whenever f is applicative functor.

newtype DayF f a

Newtype wrapper around Day f f a.

Constructors

DayF

Fields

runDayF :: Day f f a

coyonedaToDay :: forall f a. Applicative f => Coyoneda f a -> DayF f a

Note that this morphism requires f to be Applicative rather than just a Functor, like coyonedaToFreeA.

freeAToDay :: forall f a. Applicative f => FreeA f a -> DayF f a

An ismorphism with inverse dayToFreeA;

freeAToDay . dayToFreeA = id
dayToFreeA . freeAToDay = id

>>> Day.dap $ runDayF $ freeAToDay . dayToFreeA $ DayF $ Day Nothing (Just 1) (+)
Nothing
>>> Day.dap $ runDayF $ freeAToDay . dayToFreeA $ DayF $ Day (Just 1) (Just 1) (+)
Just 2
>>> foldFreeA $ dayToFreeA . freeAToDay $ FreeA Nothing (Pure 1)
Nothing
>>> foldFreeA $ dayToFreeA . freeAToDay $ FreeA (Just (+1)) (Pure 1)
Just 2

dayToFreeA :: forall f a. Applicative f => DayF f a -> FreeA f a

Left inverse of freeAToDay (split monomorhism).

Free monads

data Free f a
   = Return a
   | Free (f Free f a)

foldFree_ :: Monad m => Free m a -> m a

freeAToFree :: forall f a. Functor f => FreeA f a -> Free f a

Freer monads

newtype Freer f a

Freer monad is a composition of Free and Coyoneda. We only require that the generators f are of kind * -> * but not necessarily a functor.

Constructors

Freer

Fields

runFreer :: Free (Coyoneda f) a

foldFreer :: Monad m => Freer m a -> m a

coyonedaToFreer :: forall f a. Coyoneda f a -> Freer f a

freeToFreer :: forall f a. Functor f => Free f a -> Freer f a

Iso with inverse freeToFree, since Coyoneda f is isomorphic to f when f is a functor.

freerToFree :: forall f a. Functor f => Freer f a -> Free f a

f-algebras

class FAlgebra f a where

A uniform way of defining algebras. For example the `Algebra (FreeMonoid m) m` instance is a proof that a monoid is an `FreeMonoid-algebra` (in the sense of f-algebras). Inverse is also true, by monadicity of category of monoids. Hence having a `fold :: [m] -> m` which respects the laws of `FreeMonoid-algebras` is equivalent to having a monoid.

Minimal complete definition

fold

Methods

fold :: (FreeAlgebra f, AlgebraType f a) => f a -> a

Fold should satisfy two axioms

unitality

 gen . fold = id

associativity

                 mapFree fold
        f (f a) --------------> f a
joinFree   |                    | fold
           v                    v
         f a  ----------------> a
                   fold

f-algebras form a category with morphisms from f a -> a to f b -> b a map h :: a -> b such that

         f a -> a
 fmap h   |     | f
          v     v
         f b -> b

For all equational theories the category of f-algebras is equivalent to the cateogory of algebras for whcih f is free, e.g. category of monoids is equivalent to the category FreeMonoid-algebras, the category of semigroups is equivalent to the category of FreeSemigroup-algebras.

You can get back the algebra operations from fold:

mappend' :: Monoid m => m -> m -> m

mempty' :: Monoid m => m

Higher Kinded F-Algebras

class FAlgebra1 (m :: (* -> *) -> * -> *) (f :: * -> *) where

Minimal complete definition

fold1

Methods

fold1 :: FreeAlgebra1 m => m f a -> f a

You can get back the operations of a given algebra from fold1, e.g. for a monad, you can reconstruct both return and join (or bind):

return' :: FAlgebra1 Free m => a -> m a

join' :: forall m a. (FAlgebra1 Free m, AlgebraType1 Free m) => m (m a) -> m a

We can use the functor instance to construct join' since it is a constraint that is brought by FreeAlgebra1 Free m, we can use this constrain since it is a requirement in FreeAlgebra1 Free instance.

bind' :: forall m a b. (FAlgebra1 Free m, AlgebraType1 Free m) => m a -> (a -> m b) -> m b

For applicative functor you can get define the pure and ap

pure' :: FAlgebra1 FreeA f => a -> f a

ap' :: forall f a b. (FAlgebra1 FreeA f, AlgebraType1 FreeA f) => f (a -> b) -> f a -> f b

fmap' :: forall f a b. FAlgebra1 Coyoneda f => (a -> b) -> f a -> f b

As you can see all the operations, like <$>, <*>, join or >>= are just encoded in the data structures! This is what the free structures all about. If you read the associated post then you should connected this with the fact that free algebras are build out of terms, i.e. the operations.