Data.FreeAlgebra

Description

Author : Marcin Szamotulski, 2018 Copyright : (c) Marcin Szamotulski, 2018 License : MPL 2.0 (https:/www.mozilla.orgen-USMPL2.0/) Stability : Experimental

This is a GHC implementation of Universal Algebra approach to free algebras. For a theoretical introduction to Universal Algebra and free algebras please read "From Free Algebras to Free Monads".

Algebra type and free algebras

type family AlgebraType (a :: k) (b :: l) :: Constraint

We will use this type class to define constrains of some concrete algebras a, e.g.

instance Algebra [a] where
   type AlgebraType [a] m = Monoid m

One could also do

instance Monoid m => Algebra m where
   type AlgebraType m n = Monoid m

but one would need to deal with overlapping instances for various types of algebras: monoids ⊂ semigroups, etc; hence we will avoid it here.

class FreeAlgebra (m :: * -> *) where

A free algebra of class AlgType. An FreeAlgebra constraint proves that m is a free algebra in the class of algebras of type AlgType m over the generators gen.

The functional dependecy helps to solve otherwise ambigous types in hoistFree.

Minimal complete definition

gen, free

Methods

gen :: a -> m a

Injective map that embeds generators of m.

free

Arguments

:: AlgebraType m d

=> (a -> d)
map generators of m into d

-> m a -> d
returns a homomorphism from m to d

The freeness property.

data FreeAlgebraT m a where

Constructors

FreeAlgebraT :: m a -> FreeAlgebraT m a

foldFree :: (FreeAlgebra m, AlgebraType m a) => m a -> a

All types which satisfy FreeAlgebra constraint are foldable.

mapFree :: (FreeAlgebra m, AlgebraType m (m b)) => (a -> b) -> m a -> m b

All types which satisfy FreeAlgebra constraint are functors. The constraint AlgebraType m (m b) is always satisfied.

joinFree :: (FreeAlgebra m, AlgebraType m (m a)) => m (m a) -> m a

Further more all types which satisfy FreeAlgebra constraint are moands.

hoistFree :: forall m n a. (AlgebraType m (n a), FreeAlgebra m, FreeAlgebra n) => m a -> n a

The cannonical quotient map from a free algebra of a wider class to a free algebra of a narrower class, e.g. from free non-associative semigroup to free associative semigroup, free monoid to free commutative monoid, etc.

hoistFree . hoistFree == hoistFree

Non-Associative Semigroups

class NASemigroup a where

Non-Associative (abbrev NA) semigroup bears <> method which is not subject to any laws.

Minimal complete definition

(<>)

Methods

(<>) :: a -> a -> a

data FreeNASemigroup a

Free non-associative semigroup generated by a type a.

Constructors

Generator a

(FreeNASemigroup a) :*: (FreeNASemigroup a)

foldNA :: NASemigroup a => FreeNASemigroup a -> a

newtype Conj

Constructors

Conj

Fields

runConj :: Bool

newtype Disj

Constructors

Disj

Fields

runDisj :: Bool

Examples

type FreeNASemigroup1 = FreeNASemigroup ()

Non associative natural numbers:

(), () :*: (), (() :*: ()) :*: (), () :*: (() :*: ()), ...

type FreeNASemigroup2 = FreeNASemigroup (Sum (Succ (Succ Zero)) ())

Free non associative semigroup with two generators.

Free non-associative commutative semigroup

data Nat

Type level natural numbers promoted with DataKinds.

Constructors

Zero

Succ Nat

data Prod (n :: Nat) a where

n-fold products

Constructors

Cons :: a -> Prod n a -> Prod (Succ n) a

EmptyProduct :: Prod Zero a

p1 :: Prod (Succ (Succ (Succ Zero))) [()]

p2 :: Prod (Succ (Succ (Succ Zero))) [()]

p3 :: Prod (Succ (Succ (Succ Zero))) [()]

data Sum (n :: Nat) a where

n-fold sum

Constructors

EmptySum :: Sum Zero a

Add :: Either (Sum n a) a -> Sum (Succ n) a

s1 :: Sum (Succ n) ()

s2 :: Sum (Succ (Succ n)) ()

s3 :: Sum (Succ (Succ (Succ n))) ()

class Pointed m where

A class for pointed types, ala pointed sets, i.e. types with a distinguished term. This is one of the most simple algebras, which homomorphism preserve the distinguished element.

Minimal complete definition

mempty

Methods

mempty

Arguments

:: m
distinguished term

toProd :: forall (x :: Nat). Sum x () -> Prod x FreeMonoid1

Failed attempt to build the cannonical map from a sum to a product, we'll solve this with a ToProduct class.

class ToProduct (n :: Nat) m a b where

Minimal complete definition

toProduct

Methods

toProduct

Arguments

:: (a -> b)
how to transform coeffcients

-> m n a
a sum type

-> Prod n b
the resulting product

diagonal :: Nat -> a -> (forall n. Prod n a -> r) -> r

Construct any Prod with all values equal in CPS style.

Associative Semigroups

class NASemigroup m => Semigroup m

Class for associative semigroups, for which:

 (a <> b) <> c = a <> (b <> c)

newtype FreeSemigroup a

Free semigroup

Constructors

FreeSemigroup

Fields

runFreeSemigroup :: NonEmpty a

foldNonEmpty :: Semigroup a => FreeSemigroup a -> a

type FreeSemigroup1 = FreeSemigroup ()

Free (associative) semigroup with one generator, which is commutative.

type FreeSemigroup2 = FreeSemigroup (Either () ())

Free (associative) semigroup with two generators, hence non commutative.

toAssociative :: forall a. FreeNASemigroup a -> FreeSemigroup a

This map is the cannonical epimorphism from the free non-associative semigroup to free associative semigroup. It is uniquely determined on generators.

>>> :set -XTypeApplications
>>> runFreeSemigroup $ toAssociative $ Generator () :*: (Generator () :*: Generator ())
() :| [(),()]
>>> runFreeSemigroup $ toAssociative $ (Generator () :*: Generator ()) :*: Generator ()
() :| [(),()]

Commutative semigroups

class Semigroup m => CommutativeSemigroup m

Class of commutative semigroups, i.e. such that

a <> b = b <> a

Monoids

class (Semigroup m, Pointed m) => Monoid m

newtype FreeMonoid a

Constructors

FreeMonoid

Fields

runFreeMonoid :: [a]

foldMonoid :: Monoid m => FreeMonoid m -> m

toMonoid :: FreeSemigroup a -> FreeMonoid a

type FreeMonoid1 = FreeMonoid ()

It is isomorphic with Nat (as an additive monoid) via natToFreeMonoid1.

type FreeMonoid2 = FreeMonoid (Sum (Succ (Succ Zero)) ())

type FreeMonoid3 = FreeMonoid (Sum (Succ (Succ (Succ Zero))) ())

f1 :: FreeMonoid3

f2 :: FreeMonoid3

f3 :: FreeMonoid3

Commutative monoids

Free commutative monoid

Higher Kinded Algebras

type family AlgebraType1 (m :: k) (b :: l) :: Constraint

class FreeAlgebra1 (m :: (* -> *) -> * -> *) where

Higher kinded version of FreeAlgebra. Instances includes free functors, free applicative functors and free monads.

Minimal complete definition

gen1, free1

Methods

gen1 :: AlgebraType1 m f => f a -> m f a

Natural transformation that embeds generators into m.

free1

Arguments

:: forall (d :: * -> *). (AlgebraType m d, AlgebraType1 m f)

=> (forall x. f x -> d x)
natural transformation which embeds generators of m into d

-> m f a -> d a
a homomorphism from m to d

The freeness property.

foldFree1 :: (FreeAlgebra1 m, AlgebraType m f, AlgebraType1 m f) => m f a -> f a

hoistFree1 :: forall m n f a. (AlgebraType m (n f), AlgebraType1 m f, AlgebraType1 n f, FreeAlgebra1 m, FreeAlgebra1 n) => m f a -> n f a

Analogous to hoistFree

hoistFree1 . hoistFree1 = hoistFree1

foldCoyoneda :: Functor f => Coyoneda f a -> f a

data FreeA f a where

Free applicative functor. See Free Aplicative Functors by P.Capriotti and A.Kaposi.

Constructors

Pure :: a -> FreeA f a

FreeA :: f (x -> a) -> FreeA f x -> FreeA f a

foldFreeA :: Applicative f => FreeA f a -> f a

>>> foldFreeA (FreeA (Identity (+1)) (Pure 0))
Identity 1

coyonedaToFreeA :: forall f a. Functor f => Coyoneda f a -> FreeA f a

Day convolution

newtype DayF f a

Constructors

DayF

Fields

runDayF :: Day f f a

coyonedaToDay :: forall f a. Applicative f => Coyoneda f a -> DayF f a

Note that this morphism requires f to be Applicative rather than just a Functor, like coyonedaToFreeA.

freeAToDay :: forall f a. Applicative f => FreeA f a -> DayF f a

Split epimorphism with right inverse dayToFreeA, i.e.

freeAToDay . dayToFreeA = id

dayToFreeA :: forall f a. Applicative f => DayF f a -> FreeA f a

Left inverse of freeAToDay (split monomorhism).

Free monads

foldFree_ :: Monad m => Free m a -> m a

freeAToFree :: forall f a. Functor f => FreeA f a -> Free f a

Freer monads

newtype Freer f a

Freer monad is a composition of Free and Coyoneda. We only require that the generators f are of kind * -> * but not necessarily a functor.

Constructors

Freer

Fields

runFreer :: Free (Coyoneda f) a

foldFreer :: Monad m => Freer m a -> m a

coyonedaToFreer :: forall f a. Coyoneda f a -> Freer f a

freeToFreer :: forall f a. Functor f => Free f a -> Freer f a

Iso with inverse freeToFree, since Coyoneda f is isomorphic to f when f is a functor.

freerToFree :: forall f a. Functor f => Freer f a -> Free f a

Algebras

class Algebra f a where

A uniform way of defining algebras. For example the `Algebra (FreeMonoid m) m` instance is a proof that a monoid is an `FreeMonoid-algebra` (in the sense of f-algebras). Inverse is also true, by monadicity of category of monoids. Hence having a `fold :: [m] -> m` which respects the laws of `FreeMonoid-algebras` is equivalent to having a monoid.

Minimal complete definition

fold

Methods

fold :: (FreeAlgebra f, AlgebraType f a) => f a -> a

Higher Kinded Algebras

class Algebra1 (m :: (* -> *) -> * -> *) (f :: * -> *) where

Minimal complete definition

fold1

Methods

fold1 :: FreeAlgebra1 m => m f a -> f a